chore: Move StoreProperties to common module and added new props
