+++ components/common/src/main/java/com/bloxbean/cardano/yaci/store/common/config/StoreProperties.java
@@ -1,4 +1,4 @@
-package com.bloxbean.cardano.yaci.store.core;
+package com.bloxbean.cardano.yaci.store.common.config;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -61,4 +61,7 @@ public class StoreProperties { //TODO - replace this with YaciStoreProperties fr
     private int blocksPartitionSize=15;
     private boolean useVirtualThreadForBatchProcessing;
     private boolean useVirtualThreadForEventProcessing;
+
+    private int dbBatchSize = 200;
+    private boolean dbParallelInsert = true;
 }
