+++ components/core/src/main/java/com/bloxbean/cardano/yaci/store/core/service/CursorCleanupScheduler.java
@@ -1,6 +1,6 @@
 package com.bloxbean.cardano.yaci.store.core.service;
 
-import com.bloxbean.cardano.yaci.store.core.StoreProperties;
+import com.bloxbean.cardano.yaci.store.common.config.StoreProperties;
 import com.bloxbean.cardano.yaci.store.core.domain.Cursor;
 import com.bloxbean.cardano.yaci.store.core.storage.api.CursorStorage;
 import lombok.RequiredArgsConstructor;
