+++ components/core/src/main/java/com/bloxbean/cardano/yaci/store/core/service/publisher/ByronBlockEventPublisher.java
@@ -2,7 +2,7 @@ package com.bloxbean.cardano.yaci.store.core.service.publisher;
 
 import com.bloxbean.cardano.yaci.core.model.byron.ByronMainBlock;
 import com.bloxbean.cardano.yaci.helper.model.Transaction;
-import com.bloxbean.cardano.yaci.store.core.StoreProperties;
+import com.bloxbean.cardano.yaci.store.common.config.StoreProperties;
 import com.bloxbean.cardano.yaci.store.core.domain.Cursor;
 import com.bloxbean.cardano.yaci.store.core.service.CursorService;
 import com.bloxbean.cardano.yaci.store.events.ByronMainBlockEvent;
